+++ typings/azure-storage/azure-storage.d.ts
@@ -94,8 +94,9 @@ declare module "azure-storage" {
             * @param {string|object} [host]                       The host address. To define primary only, pass a string.
             *                                                     Otherwise 'host.primaryHost' defines the primary host and 'host.secondaryHost' defines the secondary host.
             * @param {string} [sasToken]                          The Shared Access Signature token.
+            * @param {string} [endpointSuffix]                    The endpoint suffix.
             */
-            constructor(storageAccountOrConnectionString: string, storageAccessKey?: string, host?: string|StorageHost, sasToken?: string);
+            constructor(storageAccountOrConnectionString: string, storageAccessKey?: string, host?: string|StorageHost, sasToken?: string, endpointSuffix?: string);
 
             /**
             * Associate a filtering operation with this BlobService. Filtering operations
@@ -2817,8 +2818,9 @@ declare module "azure-storage" {
           * @param {string|object} [host]                       The host address. To define primary only, pass a string. 
           *                                                     Otherwise 'host.primaryHost' defines the primary host and 'host.secondaryHost' defines the secondary host.
           * @param {string} [sasToken]                          The Shared Access Signature token.
+            * @param {string} [endpointSuffix]                    The endpoint suffix.
           */
-          constructor(storageAccountOrConnectionString?: string, storageAccessKey?: string, host?: string|StorageHost, sasToken?: string);
+          constructor(storageAccountOrConnectionString?: string, storageAccessKey?: string, host?: string|StorageHost, sasToken?: string, endpointSuffix?: string);
 
           /**
           * Associate a filtering operation with this QueueService. Filtering operations
@@ -5059,8 +5061,9 @@ declare module "azure-storage" {
           * @param {string|object} [host]                       The host address. To define primary only, pass a string.
           *                                                     Otherwise 'host.primaryHost' defines the primary host and 'host.secondaryHost' defines the secondary host.
           * @param {string} [sasToken]                          The Shared Access Signature token.
+          * @param {string} [endpointSuffix]                    The endpoint suffix.
           */
-          new (storageAccountOrConnectionString?: string, storageAccessKey?: string, host?: string|StorageHost, sasToken?: string): TableService;
+          new (storageAccountOrConnectionString?: string, storageAccessKey?: string, host?: string|StorageHost, sasToken?: string, endpointSuffix?: string): TableService;
         }
 
         export module TableUtilities {
@@ -7083,8 +7086,9 @@ declare module "azure-storage" {
           * @param {string|object} [host]                       The host address. To define primary only, pass a string. 
           *                                                     Otherwise 'host.primaryHost' defines the primary host and 'host.secondaryHost' defines the secondary host.
           * @param {string} [sasToken]                          The Shared Access Signature token.
+          * @param {string} [endpointSuffix]                    The endpoint suffix.
           */
-          new (storageAccountOrConnectionString?: string, storageAccessKey?: string, host?: string|StorageHost, sasToken?: string): FileService;
+          new (storageAccountOrConnectionString?: string, storageAccessKey?: string, host?: string|StorageHost, sasToken?: string, endpointSuffix?: string): FileService;
         }
 
         // ###########################