+++ ChangeLog.md
@@ -5,6 +5,7 @@ be taken. This is a GA release and the changes described below indicate the chan
 
 ALL
 * Fixed the issue that metadata name will be converted to lower-case after retrieving back from the server. **Note** that this fix is only applicable for Node 0.12 or higher version.
+* Added support for EndpointSuffix for all service constructors.
 
 BLOB
 * Fixed the issue that the service error message will be written to the destination stream if getting error when downloading the blob to a stream/file.